// cx-organization-spa/src/app/shared/constants/status-action-type.enum.ts
export enum StatusActionTypeEnum {
  Active = 'Active',
  Unlock = 'Unlock',
  ResetPassword = 'ResetPassword',
  Accept = 'Accept',
  Reject = 'Reject',
  CreateAccount = 'CreateAccount',
  MassCreateAccount = 'MassCreateAccount',
  SetExpirationDate = 'SetExpirationDate',
  SetApprovingOfficers = 'SetApprovingOfficers',
  Archive = 'Archive',
  RequestSpecialApproval = 'RequestSpecialApproval',
  Suspend = 'Suspend',
  AddToGroup = 'AddToGroup',
  Edit = 'Edit',
  ShowHideColumn = 'ShowHideColumn',
  Export = 'Export',
  ExportAll = 'ExportAll',
  GenerateAccountReviewReport = 'GenerateAccountReviewReport',
  ChangeUserPlaceOfWork = 'ChangeUserPlaceOfWork',
  CreateNewOrgUnit = 'CreateNewOrgUnit',
  Deactivate = 'Deactivate',
  Delete = 'Delete',
  Unarchive = 'Unarchive',
  Remove = 'Remove'
}